Add unique key check for yaml in the pack validation
