+++ st2common/st2common/util/spec_loader.py
@@ -17,21 +17,13 @@ from __future__ import absolute_import
 import pkg_resources
 
 import jinja2
-import yaml
-
-try:
-    from yaml import CSafeLoader as SafeLoader
-except ImportError:
-    from yaml import SafeLoader
-
-from yaml import constructor
-from yaml import nodes
 
 import st2common.constants.pack
 import st2common.constants.action
 
 from st2common.rbac.types import PermissionType
 from st2common.util import isotime
+from st2common.util import yml as yaml_utils
 
 __all__ = ["load_spec", "generate_spec"]
 
@@ -43,54 +35,10 @@ ARGUMENTS = {
 }
 
 
-# Custom YAML loader that throw an exception on duplicate key.
-# Credit: https://gist.github.com/pypt/94d747fe5180851196eb
-class UniqueKeyLoader(SafeLoader):
-    def construct_mapping(self, node, deep=False):
-        if not isinstance(node, nodes.MappingNode):
-            raise constructor.ConstructorError(
-                None,
-                None,
-                "expected a mapping node, but found %s" % node.id,
-                node.start_mark,
-            )
-        mapping = {}
-        for key_node, value_node in node.value:
-            key = self.construct_object(key_node, deep=deep)
-            try:
-                hash(key)
-            except TypeError as exc:
-                raise constructor.ConstructorError(
-                    "while constructing a mapping",
-                    node.start_mark,
-                    "found unacceptable key (%s)" % exc,
-                    key_node.start_mark,
-                )
-            # check for duplicate keys
-            if key in mapping:
-                raise constructor.ConstructorError(
-                    'found duplicate key "%s"' % key_node.value
-                )
-            value = self.construct_object(value_node, deep=deep)
-            mapping[key] = value
-        return mapping
-
-
-# Add UniqueKeyLoader to the yaml SafeLoader so it is invoked by safe_load.
-yaml.add_constructor(
-    yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG,
-    UniqueKeyLoader.construct_mapping,
-    Loader=SafeLoader,
-)
-
-
 def load_spec(module_name, spec_file):
     spec_string = generate_spec(module_name, spec_file)
 
-    # The use of yaml.load and passing SafeLoader is the same as yaml.safe_load which
-    # makes the same call. The difference here is that we use CSafeLoader where possible
-    # to improve performance and yaml.safe_load uses the python implementation by default.
-    return yaml.load(spec_string, SafeLoader)
+    return yaml_utils.unique_key_loader_safe_load(spec_string)
 
 
 def generate_spec(module_name, spec_file):
